Fixed problems with size_t instead of int in a loop

+++ homework/fibonacci/fibonacci.hpp
@@ -13,7 +13,7 @@ int fibonacci_iterative(int sequence) {
     int a = 0;
     int b = 1;
 
-    for (size_t i = 2; i <= sequence; ++i) {
+    for (int i = 2; i <= sequence; ++i) {
         result = a + b;
         std::swap(a, b);
         b += a;
